Better support time ranges.
However, in order to correctly determine the window we need to be aware of the
exact operator of the binary predicate.

For example, If we have an equality operator (&time == 1 day ago), the index
will never find the right segment because it only checks the bounds. Thus, for
equality we need to test against the closed interval of the segment, rather
than just a range.

Part of #1.