Implement event meta data index.
This commit implements the meta index functionality for event names and event
timestamps. Incoming queries send their query expression to the index, which in
turn checks whether it has index structures available to speed up the segment
retrieval. If it has, it sends back an index "hit". If it can, but cannot find
qualifying segments, it sends back a "miss". If it cannot speed up segment
retrieval, the index replies with an "impossible" answer, which tells the query
to ask the archive for all segment IDs and stream them sequentially.

Closes #1.