Enhance util::console.
In addition to buffer/unbuffer the console, one can now extract a single
character from STDIN with the function util::console::get. This function uses
select to poll the STDIN fd for events. It returns false either on error or
when the process received a signal, or another condition interrupted select, or
when a user-supplied timeout expires. This makes it more favorable than
std::cin::get, which blocks forever if the user does not feed it with input.