Enable bitstream polymorphism.
Eventually we will hold bitstream simply via a pointer to the base class.
We still use an intermediate basic_bitstream<T> class that serves as an
operator mixin providing bitwise operations for the direved classes. To this
end, the mixin uses the Barton-Nackman trick to inject the required binary free
functions into the global namespace.