Add storage policy for binary-encoded bitmaps.
The hash-based storage policy do not perform well in the case of binary
encoding since all log(T) bitstreams need to be accessed. The new
`vector_storage` maps values to vector indices and thus offers fast traversal,
which is well-suited when the vector is densely populated (of even full in the
case of binary encoding).