Improve bitstream and bitmap functions.
Bitmaps now support two basic primitives: push_back() and patch(). The former
just appends a value at the end of the necessary bitstream. The latter adds
invalid bits as "fill material" to the bitstream. The ability to distinguish
invalid values from the domain of the underlying type is a necessity to support
optional values. Consequently, each bitmap now has a separate bitmap that marks
the valid entries.