Use consistent expression semantics.
It's not yet clear to me how we can achieve this in an intuitive way. But at
least it's consistent now. This is the problem:

For example, should the expression ":int != +42" return T if the event does not
contain any integer, or just T when there exists an integer but not equal to
42.

Similarly, should "@1 != T" return F if the event has zero arguments? For now,
we use the inner operator (here !=) to deliver the result. That is, in both
cases the expression evaluation will return T.