Clean up signal and keyboard processing.
The system monitor no longer takes on responsibility for both signal handling
and keystroke handling. We renamed it to signal_monitor and dropped keystroke
support. We don't need keyboard interaction on the server side and client's now
can use the editline wrapper.

Moreover, we renamed util::console to util::terminal to avoid confusion with
vast::console, the new console client.