Cater to a common workaround when using "bind -v".
If users enable vi bindings via "bind -v" in their .editrc, for some reason
they experience problems with tab completion. As a result, most of them then
rebind tab to the rl_complete command. Because we do not use the readline API,
we don't have rl_complete registered and thus this fix won't work for these
guys. That's why we explicitly have to support rl_complete that users can bind
tab to it.