Provide an EWAH forward iterator.
We also changed the semantics of bitstream iterators: the dereference operation
now returns the position in the stream as opposed to just always bool. The
rationale behind this comes from the idea that one cares only about the 1-bits.
(If one wanted the 0s, then one could create the complement and iterate over
it.) However, dereference operation that would always return true is
meaningless. That's why operator* of bitstream iterators now return the
position in the bitstream. For the null bitstream, we created an iterator
adaptor that uses the 1-iterator of the bitvector. The EWAH iterator involves
more logic and stands as its own structure.

Relates #26.