Add trim() function to bitstream interface.
This function removes trailing zeros from a bitstream. For NULL bitstreams it's
a trivial function, but for EWAH bitstream it involves significantly more
logic.

This function is necessary to return minimal results when performing bitwise
operations on compressed bitstreams. That is, we cut off the result of the last
1-bit in the result.

This commit also fixes a problem in the algorithm for bitwise operation on
compressed bitstreams: after one sequence iterator reaches end, the algorithm
*always* added the remaining bits of the longer bitstream to the result. This
turns out to be incorrect, say for bitwise AND or NAND. So we added two flags
(for LHS and RHS of the operation) to control the append behavior of remaining
bits.