Add ability to rebuild index from archive.
Because indexing constitutes a very complex operation with numerous tuning
knobs, it comes in very handy to have the ability to rebuild the index from an
existing archive. This feature also decouples VAST server and client side:
rebuilding the index no longer relies on the ingestors re-feeding the raw data
to the receiver.