Let user control index partitioning.
The new command line option --index.partion allows users to manually set
partition. If it does not exist, it will be created as soon as the first
segment arrives. If it exist already, arriving segments will append to given
partition.

The index processes partitions in reverse lexicographical order. It's up to the
user to come up with sensible partition naming.