Fix bug in log parser.
Bro currently only supports sets as log output, but labels them as tables. In
VAST, a set is a record of values and the uniqueness constraint is currently
not enforced. However, it is not a table because that would be an associative
data structure. That's why we have to treat tables as records in the parser.