Support assigning event timestamp from log field.
The new command-line option --ingest.time-field takes an integer value i from
the interval  [0, n], where n represents the number of fields in the log file.
The file source then attempts to extract a timestamp at field i and assigns it
as the event timstamp. If the user does not specifiy the option on the command
line or if field i is not a timestamp type, the system's current time will be
used as event timestamp.