Implement substring search on strings.
Because the query language has currently the limitation that the LHS of a
predicate must always be an "extractor", we now have a new operator 'ni' which
has the opposite directionality of 'in'. It's essentially the set membership in
the other way. LaTeX calls it \ni as well. Moreover, we have some syntactic
sugar:

    in    ->    [+
   !in    ->    [-
    ni    ->    +]
   !ni    ->    -]

The more you squint the more [+ starts to look like the mathematical symbol for
set membership: 'e'. The minus sign is just to reflect the negation of
inclusion, which the plus signifies. Squinting does not really help here, you
won't see the vertical bar through the 'e'.