Deliver hit count as part of progress update.
Each progress update that the index pushes to the query (and the query forwards
to the client) now contains the population count of the index hit. This helps
to obtain an approximate number of the upcoming result size.

If the index lookup completes for a certain query, and the query has not a
single hit, we'll return immediate to ask mode in the console. This fixes #35.

Involves also #37.