Make mkdir more permissive in its return code.
In addition to enriching the return value of vast::mkdir, we also made it more
permissive: we no longer return a failure if the directory already exists, only
if we actually failed to create it. This is important because of TOCTTOU
problems with the file systems: if multiple actors attempt to construct the
same directory, it will no longer fail.