Support reading multiple concatenated Bro logs.
This comes in especially handy when reading goop from STDIN.

We also added an auto-detection mechanism that looks at the first value type of
a Bro log. If it is a timestamp, we use it automatically for the timestamp of
the event. It's still possible to manually select a timestamp field with the
option --ingest.time-field.