Add basic support for set/vector bitmap indexes.
This patch adds support for indexing sets/vectors with an a-priori-known value
type. That is, the element type of the container must be known at construction
time of the index. In addition, the container must be homogenous (all elements
have the same type) and not include further containers.

At this point we only allow set membership queries. When dealing with
containers, we eventually also want "any" and "all" operations coupled with
another operation basic relational operator like "<=". The set membership query
is in fact nothing but an "any" coupled with "==".

We still need to enhance the bitmap indexers so that they transparently apply a
query over the elements of a set. This is relevant for type queries, such as
":count == 42", so that they can map the existential query over all container
elements.

Relates #43.