Finish new Printable and Parseable concepts.
We have now a new printing and parsing framework. It is symmetric (The print
and parse API are a dual of each other), non-intrusive via ADL.

It is also unified as it inejcts a conversion function to<T>(..) when possible.
For example, all types T modelling Printable automatically get to_string(T) and
to<std::string>(T), where the former returns a plain string and the ladder a
trial<std::string>. Further, all types modelling the Parseable concept get a
function to<T>(std::string const&), to<T>(char const (&buf)[N]) and so forth.
This gives the function to<T>(...) a lot of power.

Further changes:

    - Move all Boost Spirit grammar definitions into the header files. While it
      increases compile times, it is necessary to do so because we cannot
      anticipate the iterator type of the parse function in advance.

    - Ignore a few warnings from Boost Spirit (Clang only).

    - Begin removal of string constructors for parseable types. Since we're not
      always able to check the input data, it is safer to go via parse() and
      then check the trial.