Shutdown gracefully when segments are in pipeline.
Now that RECEIVER processes segments synchronously, we can make it the root for
the shutdown initiation procedure. By linking ARCHIVE and INDEX to RECEIVER and
not PROGRAM, we can ensure a graceful shutdown without losing segments. This
also requires that INDEX and PARTITION do not terminate when still processing
segments. We now have this guarantee. By letting RECEIVER initiate the shutdown
process, we ensure that all acked segments will get indexed before terminating.