Avoid errors when getting strings.
Because we're using operator>> to convert a string into a given value, this
runs into problems with whitespace when attempting to get a string from a
string. E.g., if the input has "foo bar" the extracted string will only contain
"foo". We should probably dispatch as<T> for std::string, but for now we force
users to use get() instead, which simply returns the underlying string.