Enhance unit testing to actors and use CTest.
The unit test suite now includes full actor tests. At this point these simply
invoke VAST in various settings and do not perform a lot of sanity checks.
However, they lay the foundation for testing the distributed and
network-transparent aspect of VAST.

We also switched to CTest. A nice side-effect is that we can now run tests in
parallel simply by providing -j XX on the command line.

Note that 'make test' no longer ensures that all dependencies have been met.
(This is a feature and not a bug.)