Implement index flushing by means of a task tree.
The index now uses the new task tree structure to flush its bitmap indexes to
the file system. The actor initiating the flush request gets back the task
tree, which acts as a progress handle.

The task tree now operatesslightly differently: it distinguishes between a
subscriber and notifer. When subscribing to the task tree, one receives
progress for each completed edge. When asking for notification, one only gets a
done message at the end. Both types of messages are opt-in, i.e., one has to
register explicitly for them, because the simplest way of getting notification
is to monitor or link to the task tree.

Moreover, each edge has now to be removed explicitly because intermediate nodes
that spawn sub-tasks may also perform a task requiring a done message.