Major rework of fundamental type/value system.
This extensive changeset touches on the fundamental abstractions of VAST: the
notion of data, types, values, and events. The implementation of vast::value
was a hand-tuned variant structure storing the tag in one of its union members.
While space-efficient, it was neither extensible nor possible to treat
in the same style as other variant types. The new implementation uses
util::variant.

More broadly, the type system also got a big overhaul. There exists now a clear
distinction between data and types. The new class vast::data is a variant
container for data, e.g., bools, strings, sets, etc. The refactored class
vast::type a very similar interface (and also models the variant concept).

A value is data coupled with a (compatible) type. This used to be different: a
value used to be just data, but now we bring the two together to enable strong
typing throughout the entire system.

An event is-a value with a named type plus some meta data, such as timestamp
and ID. We no longer assume that events must be records values, although some
indexing code branches still have this assumption hard-coded.

Other changes:

  - Removal of vast::string in favor of std::string. We no longer work with two
    string implementations. The algorithms are now free iterator-based
    functions in vast/util/string.h.

  - New serialization code for optional<T>, enums, and variants.

  - Rework bitmap index implementation with overloads of push_back/lookup for
    the concrete types. E.g., a string_bitmap_index now provides an overload
    for std::string and not just vast::value.

  - Rename a few data structures: time_frame -> time_duration,
    prefix -> subnet, regex -> pattern.

  - New aliases for arithmetic data: integer (int64_t), real (double), and
    count (uint64_t).

  - Addition of util::arena_alloc, a stack-based allocator referencing a
    fixed-size arena. Still work-in-progress.

  - Remove util::field_splitter. The new free function util::split covers a
    superset of the functionality.

  - Tons of small fixes and new unit tests along the way.