Change CMD line semantics and source/sink factory.
VAST now sniffs packets from an interface with -i <iface> and from a
trace/stdin via -r <trace> or -r -. This is much more intuitive for most users
with a UNIX background. (Thanks for the suggestion, @rsmmr.)

Moreover, we brought the source/sink factories into the main program, where the
corresponding actor is spawned. This allows for more fine-grained and
source/sink-specific options to be passed to the endpoint.