Make 0 a valid event ID.
Up to now we used 0 as invalid event ID. This makes one thing problematic:
bitstreams have always one bit more, and we always have to ignore the invalid
position 0. When computing the complement of an index hit, this turns out to be
cumbersome and leads to annoyances in treating the exceptional case of 0.

In this commit we also consolidated the ID tracker implementation.