Improve schema parsing.
The ambiguity pointed out in 13cf7d4080 motivated a change in the syntax of the
grammar. We got rid of the Bro-style container declaration syntax and now use a
C++-style type specification. For example, a

        table[count] of set[port]

now must be written as

        table<count, set<port>>

in order to be able to unambiguously append attributes at the end of the type
specification. We also extended the grammar parser to handle attributes and
removed code referencing the now obsolete "event" keyword.

Closes #52.