Allow users to specify custom schema on import.
To control what parts of an event to index, users can now specify a schema at
import time, which may contain type annotations such as the &skip attribute to
prevent indexing. On the command line, this involves passing -s <schema-file>.

Because it can be cumbersome to come up with a default schema in the first
place, VAST also provides a "schema sniffing" capability. When specifying
--import.sniff-schema in conjunction with -I <format>, VAST will auto-deduce
the schema of the given source, print it on STDOUT, and then exit. This schema
can then be used in invocations with -s.

On export, the sinks first fetch the schema from SEARCH to ensure that
type-checks and filtering succeeds.

Closes #52.