Improve errors for non-existent dependencies.
We now keep the configuration consistent across all build types, regardless of
the presence or absence of optional dependencies. In the past, for example,
when not compiling VAST with editline then -Q would return that the
command line option would be invalid. Now VAST returns a more descriptive error
message.