Make segments an implementation detail of the archive.
We got rid of segments as they existed before. A segment used to be a sequence
of chunks, and the archive would relay entire queries to a query. The archive
now simply sends out chunks, which are much smaller and allow the query to
quicker get to the requested events. Internally, the archive still keeps a
sequence of chunks but the notion of "segment" is now an implementation detail
and no longer something users need to be aware of.

This change was necessary because the archive keeps a buffer of chunks until
they reach a certain maximum size. Then the archive would flush the chunk
sequence. But when requesting events from buffered chunks, the archive would
return with an "unknown segment" error for the request ID, because it has not
yet been registered. Now an archive lookup first goes through the list of
buffered chunks, and then hits the existing index structure to find the right
chunk. This fixes #59.

Moreover, we made the archive order-agnostic so that it can handle chunks
arriving out of order. This is an important preparation for making the whole
system more distributed, e.g., with multiple receivers, multiple importers,
multiple archives, etc.