Avoid type/data type check if schema is known.
The value constructor used to perform an implicit type check for its two
arguments data and type, even if the caller would know that the check cannot
fail. We removed this check, as it is not always desired, e.g., during import
when the schema is know by the source. If one desires a type check, there exist
now explicit factory function value::make(data, type) and event::make(value).

This change increases the mean import rate by 3.6% for Bro logs.