Make the data parser understand epoch timestamps.
The data parser could previously only parse strings like "3s5ms" but did not
understand fractional seconds, like "1398933902.686337s" as it would make sense
for a UNIX epoch time stamp. This commit adds support for fractional seconds.

Further, the time point parser had the ability to use "@X" to construct a time
point of duration X. However, this was previously not in relation to the UNIX
epoch. Now input like "@1398933902.686337s" resolves to time of 2014-05-01
08:45:02, which one would naturally expect.